+++ wiki/Changelog.md
@@ -1,5 +1,6 @@
 Desktop integration
   * WSL path conversion considers extended /etc/fstab entries (#1130).
+  * WSL path conversion supports UNC paths (#1130).
 
 ### 3.5.2 (13 November 2021) ###